Organized activities and classes into appropriate folders. Deleted unnecessary recipes.java and temp.xml.
