+++ app/src/main/AndroidManifest.xml
@@ -16,94 +16,87 @@
         android:theme="@style/Theme.TopGMeals"
         tools:targetApi="31">
         <activity
-            android:name=".AddIngredientRecipe"
+            android:name=".recipe.book.AddIngredientRecipe"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".IngredientsDisplay"
+            android:name=".recipe.book.IngredientsDisplay"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".IngredientRecipeAdapter"
+            android:name=".recipe.book.IngredientRecipe"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".IngredientRecipe"
+            android:name=".main.login.options.MainOptions"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".MainOptions"
+            android:name=".recipe.book.RecipeDisplay"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".RecipeDisplay"
+            android:name=".recipe.book.addEditRecipe"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".addEditRecipe"
+            android:name=".shopping.list.ShoppingList"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".ShoppingList"
+            android:name=".meal.plan.MealPlan"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".MealPlan"
+            android:name=".recipe.book.RecipeBook"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".RecipeBook"
-            android:exported="false">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
-            android:name=".ingredientstorage.IngredientStorage"
+            android:name=".ingredient.storage.IngredientStorage"
             android:exported="false"
             android:launchMode="singleTop"
-            android:parentActivityName=".MainActivity">
+            android:parentActivityName=".main.login.options.MainActivity">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".ingredientstorage.AddEditIngredientActivity"
+            android:name=".ingredient.storage.AddEditIngredientActivity"
             android:exported="false"
-            android:parentActivityName=".ingredientstorage.IngredientStorage">
+            android:parentActivityName=".ingredient.storage.IngredientStorage">
             <meta-data
                 android:name="android.support.IngredientStorage"
                 android:value="com/example/topgmeals/AddEditIngredientActivity.java" />
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".main.login.options.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
