--- app/src/main/java/com/example/topgmeals/ShoppingList.java
@@ -1,62 +1,0 @@
-package com.example.topgmeals;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-
-import com.example.topgmeals.ingredientstorage.IngredientStorage;
-
-public class ShoppingList extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_shopping_list);
-
-        ShoppingList currentClass = ShoppingList.this;
-        //region ButtonSwapping
-        Button IngredientButton = (Button) findViewById(R.id.switchToIngredientStorage);
-        IngredientButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(currentClass, IngredientStorage.class);
-                startActivity(intent);
-            }
-        });
-
-        Button ShoppingButton = (Button) findViewById(R.id.switchToShoppingList);
-
-        ShoppingButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(currentClass, ShoppingList.class);
-                startActivity(intent);
-            }
-        });
-
-        Button MealPlanButton = (Button) findViewById(R.id.switchToMealPlan);
-
-        MealPlanButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(currentClass, MealPlan.class);
-                startActivity(intent);
-            }
-        });
-
-        Button RecipiesButton = (Button) findViewById(R.id.switchToRecipes);
-
-        RecipiesButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(currentClass, RecipeBook.class);
-                startActivity(intent);
-            }
-        });
-        //endregion
-
-    }
-}
