+++ app/src/main/java/com/example/topgmeals/recipe/book/IngredientsDisplay.java
@@ -1,16 +1,14 @@
-package com.example.topgmeals;
+package com.example.topgmeals.recipe.book;
 
-import androidx.activity.result.ActivityResultLauncher;
-import androidx.activity.result.contract.ActivityResultContract;
-import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+
+import com.example.topgmeals.R;
 
 public class IngredientsDisplay extends AppCompatActivity {
