+++ app/src/main/java/com/example/topgmeals/recipe/book/RecipeAdapter.java
@@ -1,19 +1,18 @@
-package com.example.topgmeals;
+package com.example.topgmeals.recipe.book;
 
 import android.content.Context;
-import android.content.Intent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
-import android.widget.EditText;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.recyclerview.widget.RecyclerView;
 
-import java.util.ArrayList;
+import com.example.topgmeals.R;
+import com.example.topgmeals.recipe.book.Recipe;
+
 import java.util.List;
 
 public class RecipeAdapter extends ArrayAdapter<Recipe> {
