+++ app/src/main/java/com/example/topgmeals/shopping/list/ShoppingList.java
@@ -1,4 +1,4 @@
-package com.example.topgmeals;
+package com.example.topgmeals.shopping.list;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -7,16 +7,19 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 
-import com.example.topgmeals.ingredientstorage.IngredientStorage;
+import com.example.topgmeals.meal.plan.MealPlan;
+import com.example.topgmeals.R;
+import com.example.topgmeals.recipe.book.RecipeBook;
+import com.example.topgmeals.ingredient.storage.IngredientStorage;
 
-public class Recipes extends AppCompatActivity {
+public class ShoppingList extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_recipies);
+        setContentView(R.layout.activity_shopping_list);
 
-        Recipes currentClass = Recipes.this;
+        ShoppingList currentClass = ShoppingList.this;
         //region ButtonSwapping
         Button IngredientButton = (Button) findViewById(R.id.switchToIngredientStorage);
         IngredientButton.setOnClickListener(new View.OnClickListener() {
@@ -52,7 +55,7 @@ public class Recipes extends AppCompatActivity {
         RecipiesButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(currentClass, Recipes.class);
+                Intent intent = new Intent(currentClass, RecipeBook.class);
                 startActivity(intent);
             }
         });
