+++ app/src/main/res/layout/activity_ingredient_storage.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ingredientstorage.IngredientStorage">
+    tools:context=".ingredient.storage.IngredientStorage">
 
     <LinearLayout
         android:id="@+id/top_bar"
