+++ app/src/main/res/layout/activity_main_options.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="@drawable/option_background"
-    tools:context=".MainOptions">
+    tools:context=".main.login.options.MainOptions">
 
     <TextView
         android:id="@+id/txtWelcome"
