+++ app/src/main/res/layout/activity_meal_plan.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MealPlan">
+    tools:context=".meal.plan.MealPlan">
 
     <include layout="@layout/activity_bottomswitchbuttons"/>
