+++ app/src/main/res/layout/activity_shopping_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ShoppingList">
+    tools:context=".shopping.list.ShoppingList">
 
     <include
         android:id="@+id/include2"
