--- app/src/main/res/layout/temp.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainOptions">
-
-    <TextView
-        android:id="@+id/textView2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Welcome"
-        android:textSize="50dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        tools:layout_editor_absoluteY="74dp" />
-
-
-    <!-- This Button is the anchor for all buttons. lead mf -->
-
-    <Button
-        android:id="@+id/homeIngredientStorage"
-        android:layout_width="124dp"
-        android:layout_height="124dp"
-        android:layout_marginStart="72dp"
-        android:layout_marginTop="196dp"
-        android:text="Ingredient Storage"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <Button
-        android:id="@+id/homeShoppingList"
-        android:layout_width="124dp"
-        android:layout_height="124dp"
-        android:layout_marginStart="11dp"
-        android:text="Shopping List"
-        app:layout_constraintBaseline_toBaselineOf="@+id/homeIngredientStorage"
-        app:layout_constraintStart_toEndOf="@+id/homeIngredientStorage" />
-
-    <Button
-        android:id="@+id/homeMealPlanner"
-        android:layout_width="124dp"
-        android:layout_height="124dp"
-        android:layout_marginTop="13dp"
-        android:text="Meal Planner"
-        app:layout_constraintStart_toStartOf="@+id/homeIngredientStorage"
-        app:layout_constraintTop_toBottomOf="@+id/homeIngredientStorage" />
-
-    <Button
-        android:id="@+id/homeRecipesBook"
-        android:layout_width="124dp"
-        android:layout_height="124dp"
-        android:layout_marginStart="8dp"
-        android:text="Recipes Book"
-        app:layout_constraintBaseline_toBaselineOf="@+id/homeMealPlanner"
-        app:layout_constraintStart_toEndOf="@+id/homeMealPlanner" />
-</androidx.constraintlayout.widget.ConstraintLayout>
